windows_ports: Required code for windows dll to build Resource file for windows dll versioning various fixes for warnings given by visual studio compiler

+++ EZ_SDK/PWM.cpp
@@ -15,7 +15,7 @@ int PWMClass::GetPWM(DigitalClass::DigitalPortEnum pwmPort){
 }
 
 bool PWMClass::IsPWMStopped(DigitalClass::DigitalPortEnum pwmPort){
-	return m_pwmReleased[(int) pwmPort];
+	return m_pwmReleased[pwmPort];
 }
 
 void PWMClass::SetPWM(DigitalClass::DigitalPortEnum pwmPort, int position)
